Add files via upload
Added functions replacing previous drawBar, drawPie, drawPixel and drawCircle. Also added colors defined with Color class. The only thing left to do is to change putImage and getImage functions in WINDOW.cpp. The drawing functions are defined in Gate.h class. There is a function drawImage, which should be used instead of the older putImage.